Move Calendar from DateInput/index.tsx to its own file DateInput/Calendar.tsx
